requirements.txt
torch==1.6.0
torchvision==0.7.0
streamlit==0.64.0
numpy==1.22.0
Pillow==9.0.1
